README.md update, code changes to Chat og MicroMenu.
Extensive list of what's new in this version of Nurfed, and what's
missing.

MicroMenu.lua
Made the arrays for the menus more readable, and added some comments.

Chat.lua
Changed the name of a variable to be more understandable. Updated the
references to this variable.

Nurfed.lua
Added 2 comments only.